+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.46] - 2023-06-16
+### Added
+* description containers to `PrefabManager` allowing for descriptive bundling of Prefabs
+* optional parameters to `IPoolable`'s `static` `Create()` and `PoolSystem`'s `CreateIPoolable()`
+### Fixed
+* incorrect optional parameter handling within `IPoolable` and `PoolSystem`
+
 ## [0.0.45] - 2023-06-15
 ### Added
 * slight optimization to `ListPool`